Feed create and index done

+++ database/migrations/2021_04_16_184038_create_ingredients_table.php
@@ -17,8 +17,8 @@ class CreateIngredientsTable extends Migration
             $table->foreignId('raw_id')->constrained()->onDelete('cascade');
             $table->foreignId('supplier_id')->constrained();
             $table->string('unit');
-            $table->float('unit_price');
-            $table->float('amount');
+            $table->decimal('unit_price',6, 2);
+            $table->decimal('amount', 6, 2);
             $table->string('file')->nullable();
             $table->string('qc_report');
             $table->boolean('is_approved')->default(false);
